fix Codec Default implementation
