+++ src/main/java/com/alipay/remoting/RemotingServer.java
@@ -30,6 +30,7 @@ public interface RemotingServer {
      */
     boolean start();
 
+
     /**
      * Stop the server.
      *
